pyproject.toml
[project]
name = "invoice-extraction-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google>=3.0.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-documentai>=3.5.0",
    "google-cloud-documentai-toolbox>=0.14.2a0",
    "google-generativeai>=0.8.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
